Rename Numeric to Numerics, fix includes
