+++ src/Core/CMakeLists.txt
@@ -6,7 +6,7 @@ set(public_files
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/FileSystem.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/LocoFixedVector.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/MemoryStream.h"
-    "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/Numeric.hpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/Numerics.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/Prng.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/Span.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/include/OpenLoco/Core/Stream.hpp"
@@ -17,7 +17,7 @@ set(private_files
     "${CMAKE_CURRENT_SOURCE_DIR}/src/BinaryStream.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/FileStream.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/MemoryStream.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Numeric.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Numerics.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Prng.cpp"
 )
 
@@ -25,7 +25,7 @@ set(test_files
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/EnumFlagsTests.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/FileStreamTests.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/MemoryStreamTests.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/tests/NumericTests.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/tests/NumericsTests.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/PrngTests.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/tests/TraitsTest.cpp"
 )
