+++ src/Core/include/OpenLoco/Core/Numerics.hpp
