+++ src/Core/src/Numerics.cpp
@@ -1,4 +1,4 @@
-#include "Numeric.hpp"
+#include "Numerics.hpp"
 
 #include <cstddef>
 #include <cstdint>
