+++ src/OpenLoco/src/Windows/CompanyList.cpp
@@ -18,7 +18,7 @@
 #include "World/Company.h"
 #include "World/CompanyManager.h"
 #include "World/CompanyRecords.h"
-#include <OpenLoco/Core/Numeric.hpp>
+#include <OpenLoco/Core/Numerics.hpp>
 #include <OpenLoco/Interop/Interop.hpp>
 
 using namespace OpenLoco::Interop;
