+++ src/OpenLoco/src/Windows/TownList.cpp
@@ -22,7 +22,7 @@
 #include "Widget.h"
 #include "World/Town.h"
 #include "World/TownManager.h"
-#include <OpenLoco/Core/Numeric.hpp>
+#include <OpenLoco/Core/Numerics.hpp>
 #include <OpenLoco/Interop/Interop.hpp>
 
 using namespace OpenLoco::Interop;
