+++ src/OpenLoco/src/World/Town.cpp
@@ -11,7 +11,7 @@
 #include "Random.h"
 #include "TownManager.h"
 #include "Ui/WindowManager.h"
-#include <OpenLoco/Core/Numeric.hpp>
+#include <OpenLoco/Core/Numerics.hpp>
 #include <OpenLoco/Interop/Interop.hpp>
 #include <algorithm>
